Seperated the auto start component
